+++ src/Taskling/Fluent/IFluentBlockSettingsDescriptor.cs
@@ -4,7 +4,7 @@ namespace Taskling.Fluent;
 
 public interface IFluentBlockSettingsDescriptor
 {
-    IFluentBlockSettingsDescriptor ReprocessFailedTasks(TimeSpan detectionRange, short retryLimit);
-    IFluentBlockSettingsDescriptor ReprocessDeadTasks(TimeSpan detectionRange, short retryLimit);
+    IFluentBlockSettingsDescriptor ReprocessFailedTasks(TimeSpan detectionRange, int retryLimit);
+    IFluentBlockSettingsDescriptor ReprocessDeadTasks(TimeSpan detectionRange, int retryLimit);
     IComplete MaximumBlocksToGenerate(int maximumNumberOfBlocks);
 }
