+++ src/Taskling/Fluent/ListBlocks/IFluentListBlockDescriptorBase.cs
@@ -5,9 +5,9 @@ namespace Taskling.Fluent;
 
 public interface IFluentListBlockDescriptorBase<T>
 {
-    IOverrideConfigurationDescriptor WithSingleUnitCommit(List<T> values, short maxBlockSize);
-    IOverrideConfigurationDescriptor WithPeriodicCommit(List<T> values, short maxBlockSize, BatchSize batchSize);
-    IOverrideConfigurationDescriptor WithBatchCommitAtEnd(List<T> values, short maxBlockSize);
+    IOverrideConfigurationDescriptor WithSingleUnitCommit(List<T> values, int maxBlockSize);
+    IOverrideConfigurationDescriptor WithPeriodicCommit(List<T> values, int maxBlockSize, BatchSize batchSize);
+    IOverrideConfigurationDescriptor WithBatchCommitAtEnd(List<T> values, int maxBlockSize);
     IReprocessScopeDescriptor ReprocessWithSingleUnitCommit();
     IReprocessScopeDescriptor ReprocessWithPeriodicCommit(BatchSize batchSize);
     IReprocessScopeDescriptor ReprocessWithBatchCommitAtEnd();
@@ -15,12 +15,12 @@ public interface IFluentListBlockDescriptorBase<T>
 
 public interface IFluentListBlockDescriptorBase<TItem, THeader>
 {
-    IOverrideConfigurationDescriptor WithSingleUnitCommit(List<TItem> values, THeader header, short maxBlockSize);
+    IOverrideConfigurationDescriptor WithSingleUnitCommit(List<TItem> values, THeader header, int maxBlockSize);
 
-    IOverrideConfigurationDescriptor WithPeriodicCommit(List<TItem> values, THeader header, short maxBlockSize,
+    IOverrideConfigurationDescriptor WithPeriodicCommit(List<TItem> values, THeader header, int maxBlockSize,
         BatchSize batchSize);
 
-    IOverrideConfigurationDescriptor WithBatchCommitAtEnd(List<TItem> values, THeader header, short maxBlockSize);
+    IOverrideConfigurationDescriptor WithBatchCommitAtEnd(List<TItem> values, THeader header, int maxBlockSize);
     IReprocessScopeDescriptor ReprocessWithSingleUnitCommit();
     IReprocessScopeDescriptor ReprocessWithPeriodicCommit(BatchSize batchSize);
     IReprocessScopeDescriptor ReprocessWithBatchCommitAtEnd();
